requirements.txt
streamlit
easyocr
pytesseract
opencv-python
numpy
pillow
sympy
